moving getPositiveRemaindersOrFail method to TxUtils trait
